+++ src/Velvet.WebGL/ts/core/buffers/IBuffer.ts
@@ -1,6 +1,6 @@
 export interface IBuffer {
     id: number;
-    setData(data: Float32Array | Uint16Array): void;
+    setData(data: Float32Array | Uint32Array): void;
     bind(): void;
     unbind(): void;
     delete(): void;
